jobs/requirements.txt
# Python requirements for XandTube Channel Tracking Jobs
# Install with: pip install -r requirements.txt

# Core scheduling
APScheduler>=3.10.4

# Database connectivity
psycopg2-binary>=2.9.7

# HTTP requests
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Date/time handling
pytz>=2023.3

# Logging
colorlog>=6.7.0

# JSON handling (built-in, but explicit for clarity)
# json - built-in

# System operations
# os - built-in
# subprocess - built-in
# datetime - built-in

# Optional: Better error handling and retries
tenacity>=8.2.3